+++ packages/core-services/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocket.chat/core-services
 
+## 0.3.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.3
+- @rocket.chat/rest-typings@6.5.3
+- @rocket.chat/models@0.0.27
+
 ## 0.3.2
 
 ### Patch Changes